web/requirements/common.txt
Django==2.0.1
django-rest-knox==3.1.2
djangorestframework==3.7.7
psycopg2==2.7.3.2
gunicorn==19.7.1